Remove duplications by extracting common code into a generic PageProvider
